third_party/WebKit/LayoutTests/inspector/console/console-xml-document-expected.txt

CONSOLE MESSAGE: line 9: [object XMLDocument]
CONSOLE MESSAGE: line 11: [object Element]
Tests that XML document contents are logged using the correct case in the console.

console-xml-document.html:4 #document<MixedCase> Test </MixedCase>
console-xml-document.html:6 <Book></Book>
